Add backtracking linesearch to GradientMethod
Backtracking linesearch was implemented following the FISTA paper,
and supported with and without proximal, and with and without
acceleration. Linesearch is useful when the Lipschitz constant is
unknown or cannot be calculated efficiently. For regularized least
squares problems, linesearch may not be as useful because the
Lipschitz constant can be estimated using the power method.